requirements.txt
azure-functions==1.17.0
fastapi==0.104.1
uvicorn==0.24.0.post1